Some tidy-up, hope it still works. I did comment out some stuff to make the code compile.
