+++ src/get_coocurrence_vector.cpp
@@ -1,3 +1,4 @@
+#include "get_coocurrence_matrix.h"
 #define DARMA_64BIT_WORD 1
 #include <RcppArmadillo.h>
 using namespace Rcpp;
@@ -8,49 +9,6 @@ using namespace Rcpp;
 
 // [[Rcpp::depends(RcppArmadillo)]]
 // [[Rcpp::plugins(cpp11)]]
-
-// [[Rcpp::export]]
-IntegerMatrix rcpp_get_coocurrence_matrix(arma::imat x, arma::imat directions) {
-
-    // get unique values
-    arma::ivec u = arma::conv_to<arma::ivec>::from(arma::unique(x.elem(find(x != INT_MIN))));
-
-    // create a matrix of zeros of unique values size
-    arma::imat cooc_mat(u.n_elem, u.n_elem, arma::fill::zeros);
-
-    // extract adjency pairs
-    List pairs = rcpp_get_pairs(x, directions);
-
-    IntegerVector center_cells = pairs[0];
-    IntegerVector neigh_cells = pairs[1];
-
-    // number of pairs
-    int num_pairs = center_cells.length();
-
-    // for each row and col
-    for (int i = 0; i < num_pairs; i++) {
-
-        // extract value of central cell and its neighbot
-        int center = center_cells(i);
-        int neigh = neigh_cells(i);
-
-        // find their location in the output matrix
-        arma::uvec loc_c = find(u == center);
-        arma::uvec loc_n = find(u == neigh);
-
-        // add its count
-        cooc_mat(loc_c, loc_n) += 1;
-
-    }
-
-    // return a coocurence matrix
-    IntegerMatrix cooc_mat_result = as<IntegerMatrix>(wrap(cooc_mat));
-
-    // add names
-    List u_names = List::create(u, u);
-    cooc_mat_result.attr("dimnames") = u_names;
-    return cooc_mat_result;
-}
 
 // [[Rcpp::export]]
 int triangular_index(int r, int c) {
@@ -106,13 +64,14 @@ NumericVector rcpp_get_offdiagonal_vector(arma::imat x, arma::imat directions) {
 }
 
 // [[Rcpp::export]]
+/*
 NumericVector moving_filter(arma::imat x, arma::imat directions) {
 
     // get unique values
     arma::ivec u = arma::conv_to<arma::ivec>::from(arma::unique(x.elem(find(x != INT_MIN))));
 
     // create empty cooccurence matrix
-    arma::imat cooc_mat(u,u)
+    arma::imat cooc_mat(u,u);
 
     // get number of rows and cols for the looping
     int nrows = x.n_rows();
@@ -145,7 +104,7 @@ NumericVector moving_filter(arma::imat x, arma::imat directions) {
 
     }
 }
-
+*/
 /*** R
 
 */
